make clippy happy

+++ nativeshell/src/shell/platform/win32/hot_key.rs
@@ -86,7 +86,7 @@ impl PlatformHotKeyManager {
         self.hot_keys.borrow_mut().insert(handle, engine);
         unsafe {
             let vk = MapVirtualKeyW(virtual_key as u32, MAPVK_VSC_TO_VK);
-            RegisterHotKey(self.hwnd(), handle.0 as i32, modifiers, vk as u32);
+            RegisterHotKey(self.hwnd(), handle.0 as i32, modifiers, vk);
         }
         Ok(())
     }
